Properly set up higher half, moved kernel even higher. Feral was relying upon a weird quirk with QEMU that was causing it to work when a specific address was used for relocation.
That's been changed, and now Feral loads at an
even higher offset, so more room can be dedicated
to usermode.

In addition, some minor refactoring was done
to for more assembly source files, however,
this requires some editing in the arch/*
makefiles to be done, since nasm
doesn't like having multiple inputs to deal with.

Feral *should* work on some other emulators,
and I'll have to go look for some time
to go test on a real PC soon.