Initial heap implementation Untested, but it seems like it should work whenever it gets used.
Specifically, this is intended to scale well,
be able to be improved relatively easilly,
and be suitable as-is for a "real" system.

Some things to consider are making the
nodes for the heap into a skip list,
so that it's possible to just skip many
nodes if the malloc size if large enough.
This still needs to be tested though: it compiles,
but the kernel never actually uses it, yet.